requirements.txt
alembic==1.7.3
autopep8==1.5.7
blinker==1.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
Pillow==8.3.2
pycodestyle==2.7.0
SQLAlchemy==1.4.25
toml==0.10.2
visitor==0.1.3
Werkzeug==2.0.1
Whoosh==2.7.4
WTForms==2.3.3
zipp==3.5.0
